Separating views.py to files, kubernetes configuration
Moved functions from views.py to users_views.py. Changed __init__.py for views. Added filebeat configuration files.
